修改了几处代码，并对文件进行调整
