+++ qulab_toolbox/wavedata/_vIQmixer.py
@@ -1,6 +1,6 @@
 import numpy as np
-from ._wavedata import Wavedata,Sin,Cos
-from ._wd_func import Exp
+from ._wavedata import Wavedata
+from ._wd_func import Exp,Sin,Cos
 
 '''Wavedata 虚拟IQ混频器模块'''
 
@@ -84,7 +84,7 @@ class vIQmixer(object):
         return vIQ._RF
 
     @classmethod
-    def carry_wave(cls,carry_freq,I=0,Q=0,IQ=None,carry_cali=None,DEG=True):
+    def carry_wave(cls,carry_freq=0,I=0,Q=0,IQ=None,carry_cali=None,DEG=True):
         '''将I/Q分别加载某个频率的载波，
         carry_cali对应实体IQ混频器的校准矩阵，与上面cali_array格式相同'''
         if carry_cali is None:
@@ -105,8 +105,9 @@ class vIQmixer(object):
         _carry_IQ = IQ*Exp(2*np.pi*carry_freq,0,IQ.len,IQ.sRate)
 
         # 相位校准，等效于进行波形时移，时移大小由相位误差、频率等决定
-        shift_I = _phi_I/(2*np.pi*carry_freq)
-        shift_Q = _phi_Q/(2*np.pi*carry_freq)
+        # 如果载波频率为0，则不进行相位校准
+        shift_I = _phi_I/(2*np.pi*carry_freq) if not carry_freq==0 else 0
+        shift_Q = _phi_Q/(2*np.pi*carry_freq) if not carry_freq==0 else 0
 
         # I/Q分别进行校准
         carry_I = (_carry_IQ.I()<<shift_I)*_scale_I+_offset_I
